requirements.txt
numpy
opencv-python
Pillow
tensorflow
mysql-connector-python
matplotlib